# perf(inductor): improve `Adam` compile times by shortcutting for loops (via `has_complex`)