--- a/torch/distributed/optim/functional_adam.py
+++ b/torch/distributed/optim/functional_adam.py
@@ -73,6 +73,7 @@ class _FunctionalAdam:
         exp_avg_sqs = []
         max_exp_avg_sqs = []
         state_steps: List[Tensor] = []
+        has_complex = torch.is_complex(param)
         if grad is not None:
             params_with_grad.append(param)
             grads.append(grad)
@@ -108,6 +109,7 @@ class _FunctionalAdam:
                 max_exp_avg_sqs,
                 state_steps,
                 amsgrad=self.amsgrad,
+                has_complex=has_complex,
                 maximize=self.maximize,
                 beta1=self.defaults["beta1"],
                 beta2=self.defaults["beta2"],
@@ -128,6 +130,7 @@ class _FunctionalAdam:
         exp_avg_sqs = []
         max_exp_avg_sqs = []
         state_steps: List[Tensor] = []
+        has_complex = False
 
         if len(params) != len(gradients):
             raise ValueError(
@@ -138,6 +141,8 @@ class _FunctionalAdam:
 
         for param, gradient in zip(self.param_group["params"], gradients):
             if gradient is not None:
+                if torch.is_complex(param):
+                    has_complex = True
                 params_with_grad.append(param)
                 grads.append(gradient)
                 # Lazy state initialization
@@ -178,6 +183,7 @@ class _FunctionalAdam:
                 max_exp_avg_sqs,
                 state_steps,
                 amsgrad=self.amsgrad,
+                has_complex=has_complex,
                 maximize=self.maximize,
                 beta1=self.defaults["beta1"],
                 beta2=self.defaults["beta2"],