--- a/torch/optim/adam.py
+++ b/torch/optim/adam.py
@@ -592,7 +592,7 @@ def _fused_adam(
     found_inf: Optional[Tensor],
     *,
     amsgrad: bool,
-    has_complex: bool, # Needed for consistency.
+    has_complex: bool,  # Needed for consistency.
     beta1: float,
     beta2: float,
     lr: Union[float, Tensor],