@janeyx99 2023-10-05
Ah I wasn't sure if the "has_complex" referred to the combination or just the has_complex flag, but I should also get better at reading 😅 A 2X dynamo speedup for num_params=1k seems worth doing, especially when I expect the majority of use cases to not use complex numbers.